Prevent_unsafe_narrowing: base/ for Linux part 3 of n
Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3740189
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020888}

+++ base/cpu.h
@@ -130,7 +130,7 @@ class BASE_EXPORT CPU final {
 
   struct TimeInStateEntry {
     CPU::CoreType core_type;      // type of the cores in this cluster.
-    uint32_t cluster_core_index;  // index of the first core in the cluster.
+    size_t cluster_core_index;    // index of the first core in the cluster.
     uint64_t core_frequency_khz;
     TimeDelta cumulative_time;
   };
