+++ base/memory/page_size_posix.cc
@@ -13,7 +13,7 @@ size_t GetPageSize() {
   // For more information see getpagesize(2). Portable applications should use
   // sysconf(_SC_PAGESIZE) rather than getpagesize() if it's available.
 #if defined(_SC_PAGESIZE)
-    return sysconf(_SC_PAGESIZE);
+    return static_cast<size_t>(sysconf(_SC_PAGESIZE));
 #else
     return getpagesize();
 #endif
